Patch connection event defaults for improved performance/reliability.
The NimBLE controller does not make use of the Minimum and Maximum Connection Event lengths
but the ESP32 controller does. The NimBLE default values create error conditions in the ESP32
controller when connecting to multiple peers with connection intervals.

This patch sets those parameters to 0 to allow the controller to manage the time window.