Revert 42201d4, app should specify response when (un)subscribing.
In some applications it is benefitial to not write to the 0x2902 descriptor with response, particulary in cases where the device sends
notifications regardless of subscribing. This reverts the previous change to the behavior of the (un)subscribe call which forced writing
with response, the application should decide if it's required.