Revert e75b406 sub/unsub changes, app should specify response.
In some applications it is beneficial to not write to the 0x2902 descriptor with response, particularly in cases where the device sends notifications regardless of subscribing. This reverts the previous change to the behavior of the (un)subscribe call which forced writing with response, the application should decide if it's required.